+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 # Changelog
+## Jul 29, 2024
+- Update Dashboard: Display Workout Routines, Activities Logged.
+
 ## Jul 28, 2024
 - Updated Frontend: Added `Create Workout Routine` Component.
 - Updated Frontend: Added `Activity` Component.